Move NaviteSoapClient
